pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sheetbridge"
version = "0.1.0"
description = "Google Sheets client wrapping gspread with DataFrame sync and change detection"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "Josh Wren" }]
dependencies = [
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "pandas>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["sheetbridge"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=sheetbridge",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=96",
]

[tool.coverage.run]
source = ["sheetbridge"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "D", "ANN401"]
ignore = ["D203", "D213", "E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
include = ["sheetbridge", "tests"]
exclude = ["**/node_modules", "**/__pycache__", ".venv"]
venvPath = "."
venv = ".venv"
